+++ main/heroes/Knight.java
@@ -3,7 +3,7 @@ package main.heroes;
 public class Knight extends Player {
 //    public final String name;
     private int bonusHpPerLevel = 80;
-    private double initialHp = 900;
+    private float initialHp = 900;
     public Knight(int x, int y) {
         this.hp = initialHp;
         this.x = x;
@@ -13,8 +13,8 @@ public class Knight extends Player {
     }
 
     @Override
-    public double isAttackedBy(Player player, char landType) {
-        double firstDamage = player.attack(this, landType);
+    public float isAttackedBy(Player player, char landType) {
+        float firstDamage = player.attack(this, landType);
 //        System.out.println(player.getLevel());
 //        System.out.println(player.getName());
 //        System.out.println("Tipul pamantului: " + landType);
@@ -24,12 +24,12 @@ public class Knight extends Player {
     }
 
     @Override
-    public double attack(Knight knight, char landType) {
-        double damage;
+    public float attack(Knight knight, char landType) {
+        float damage;
         damage = Constants.execute + Constants.executeBonus * knight.getLevel();
-        double hpLimit;
-        double teoreticMaxHp = knight.getHp() + Constants.knightHp + Constants.knightHpPerLevel * knight.getLevel();
-        double maxProcent = 0.01f * knight.getLevel();
+        float hpLimit;
+        float teoreticMaxHp = knight.getHp() + Constants.knightHp + Constants.knightHpPerLevel * knight.getLevel();
+        float maxProcent = 0.01f * knight.getLevel();
         if (maxProcent > 0.4f) {
             maxProcent = 0.4f;
         }
@@ -39,7 +39,7 @@ public class Knight extends Player {
             return damage;
         }
         damage *= Constants.executeKnight;
-        double slam;
+        float slam;
         knight.canMove = false;
         knight.overtimeDamage = 0;
         knight.overtimeRounds = 1;
@@ -53,12 +53,12 @@ public class Knight extends Player {
     }
 
     @Override
-    public double attack(Pyromancer pyromancer, char landType) {
-        double damage;
+    public float attack(Pyromancer pyromancer, char landType) {
+        float damage;
         damage = Constants.execute + Constants.executeBonus * pyromancer.getLevel();
-        double hpLimit;
-        double teoreticMaxHp = pyromancer.getHp() + Constants.knightHp + Constants.knightHpPerLevel * pyromancer.getLevel();
-        double maxProcent = 0.01f * pyromancer.getLevel();
+        float hpLimit;
+        float teoreticMaxHp = pyromancer.getHp() + Constants.knightHp + Constants.knightHpPerLevel * pyromancer.getLevel();
+        float maxProcent = 0.01f * pyromancer.getLevel();
         if (maxProcent > 0.4f) {
             maxProcent = 0.4f;
         }
@@ -68,7 +68,7 @@ public class Knight extends Player {
             return damage;
         }
         damage *= Constants.executePyromancer;
-        double slam;
+        float slam;
         pyromancer.canMove = false;
         pyromancer.overtimeDamage = 0;
         pyromancer.overtimeRounds = 1;
@@ -82,12 +82,12 @@ public class Knight extends Player {
     }
 
     @Override
-    public double attack(Rogue rogue, char landType) {
-        double damage;
+    public float attack(Rogue rogue, char landType) {
+        float damage;
         damage = Constants.execute + Constants.executeBonus * rogue.getLevel();
-        double hpLimit;
-        double teoreticMaxHp = rogue.getHp() + Constants.knightHp + Constants.knightHpPerLevel * rogue.getLevel();
-        double maxProcent = 0.01f * rogue.getLevel();
+        float hpLimit;
+        float teoreticMaxHp = rogue.getHp() + Constants.knightHp + Constants.knightHpPerLevel * rogue.getLevel();
+        float maxProcent = 0.01f * rogue.getLevel();
         if (maxProcent > 0.4f) {
             maxProcent = 0.4f;
         }
@@ -99,7 +99,7 @@ public class Knight extends Player {
         }
 
         damage *= Constants.executeRogue;
-        double slam;
+        float slam;
         rogue.canMove = false;
         rogue.overtimeDamage = 0;
         rogue.overtimeRounds = 1;
@@ -114,12 +114,12 @@ public class Knight extends Player {
     }
 
     @Override
-    public double attack(Wizard wizard, char landType) {
-        double damage;
+    public float attack(Wizard wizard, char landType) {
+        float damage;
         damage = Constants.execute + Constants.executeBonus * wizard.getLevel();
-        double hpLimit;
-        double teoreticMaxHp = wizard.getHp() + Constants.knightHp + Constants.knightHpPerLevel * wizard.getLevel();
-        double maxProcent = 0.01f * wizard.getLevel();
+        float hpLimit;
+        float teoreticMaxHp = wizard.getHp() + Constants.knightHp + Constants.knightHpPerLevel * wizard.getLevel();
+        float maxProcent = 0.01f * wizard.getLevel();
         if (maxProcent > 0.4f) {
             maxProcent = 0.4f;
         }
@@ -129,7 +129,7 @@ public class Knight extends Player {
             return damage;
         }
         damage *= Constants.executeWizard;
-        double slam;
+        float slam;
         wizard.canMove = false;
         wizard.overtimeDamage = 0;
         wizard.overtimeRounds = 1;
