+++ main/heroes/Pyromancer.java
@@ -2,7 +2,7 @@ package main.heroes;
 
 public class Pyromancer extends Player {
 //    public final String name;
-    private double initialHp = 500;
+    private float initialHp = 500;
     private int bonusHpPerLevel = 50;
     private int bonusDamagePerLevel = 50;
     private int currentRound;
@@ -17,17 +17,17 @@ public class Pyromancer extends Player {
 
 
     @Override
-    public double isAttackedBy(Player player, char landType) {
-        double damage = player.attack(this, landType);
+    public float isAttackedBy(Player player, char landType) {
+        float damage = player.attack(this, landType);
         damage = Math.round(damage);
         return damage;
     }
     @Override
-    public double attack(Knight knight, char landType) {
-        double damage;
+    public float attack(Knight knight, char landType) {
+        float damage;
         damage = Constants.fireblast + knight.getLevel() * Constants.fireblastBonus;
         damage *= Constants.fireblastKnight;
-        double igniteDamage;
+        float igniteDamage;
         igniteDamage = Constants.ignite + knight.getLevel() * Constants.igniteBonus;
         knight.overtimeRounds = 2;
         knight.overtimeDamage = Constants.igniteOvertime + (Constants.igniteOvertimeBonus * knight.getLevel());
@@ -43,11 +43,11 @@ public class Pyromancer extends Player {
     }
 
     @Override
-    public double attack(Pyromancer pyromancer, char landType) {
-        double damage;
+    public float attack(Pyromancer pyromancer, char landType) {
+        float damage;
         damage = Constants.fireblast + pyromancer.getLevel() * Constants.fireblastBonus;
         damage *= Constants.fireblastPyromancer;
-        double igniteDamage;
+        float igniteDamage;
         igniteDamage = Constants.ignite + pyromancer.getLevel() * Constants.igniteBonus;
         pyromancer.overtimeRounds = 2;
         pyromancer.overtimeDamage = Constants.igniteOvertime + (Constants.igniteOvertimeBonus * pyromancer.getLevel());
@@ -63,11 +63,11 @@ public class Pyromancer extends Player {
     }
 
     @Override
-    public double attack(Rogue rogue, char landType) {
-        double damage;
+    public float attack(Rogue rogue, char landType) {
+        float damage;
         damage = Constants.fireblast + rogue.getLevel() * Constants.fireblastBonus;
         damage *= Constants.fireblastRogue;
-        double igniteDamage;
+        float igniteDamage;
         igniteDamage = Constants.ignite + rogue.getLevel() * Constants.igniteBonus;
         rogue.overtimeRounds = 2;
         rogue.overtimeDamage = Constants.igniteOvertime + (Constants.igniteOvertimeBonus * rogue.getLevel());
@@ -83,11 +83,11 @@ public class Pyromancer extends Player {
     }
 
     @Override
-    public double attack(Wizard wizard, char landType) {
-        double damage;
+    public float attack(Wizard wizard, char landType) {
+        float damage;
         damage = Constants.fireblast + wizard.getLevel() * Constants.fireblastBonus;
         damage *= Constants.fireblastWizard;
-        double igniteDamage;
+        float igniteDamage;
         igniteDamage = Constants.ignite + wizard.getLevel() * Constants.igniteBonus;
         wizard.overtimeRounds = 2;
         wizard.overtimeDamage = Constants.igniteOvertime + (Constants.igniteOvertimeBonus * wizard.getLevel());
